requirements.txt
Django
djangorestframework
djoser
sqlparse
Pillow
gunicorn
psycopg2